+++ include/mm2/rx/decoder.hpp
@@ -10,23 +10,21 @@
 
 #pragma once
 
+#include <cstdint>
 #include "readable.hpp"
 #include "writable.hpp"
 
 namespace mm2::rx {
 
 template<typename T>
-concept Decoder = Readable<T> && Writable<T> &&
-                  requires(T t,
-                           uint32_t addr,
-                           bool dir,
-                           int32_t speed,
-                           uint32_t mask,
-                           uint32_t state) {
-                    { t.direction(addr, dir) } -> std::same_as<void>;
-                    { t.speed(addr, speed) } -> std::same_as<void>;
-                    { t.reverse(addr) } -> std::same_as<void>;
-                    { t.function(addr, mask, state) } -> std::same_as<void>;
-                  };
+concept Decoder =
+  Readable<T> && Writable<T> &&
+  requires(
+    T t, uint8_t addr, bool dir, int32_t speed, uint32_t mask, uint32_t state) {
+    { t.direction(addr, dir) } -> std::same_as<void>;
+    { t.speed(addr, speed) } -> std::same_as<void>;
+    { t.reverse(addr) } -> std::same_as<void>;
+    { t.function(addr, mask, state) } -> std::same_as<void>;
+  };
 
 } // namespace mm2::rx
